requirements.txt
transformers>=4.43.0
accelerate>=0.33.0
torch>=2.2.0
torchvision>=0.17.0
numpy>=1.26.0
pandas>=2.2.0
pillow>=10.2.0
tqdm>=4.66.0
statsmodels>=0.14.1
scipy>=1.11.0
